requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
python-dotenv==1.0.0
et-xmlfile==1.1.0
mysqlclient==2.2.4
openpyxl==3.1.5
pillow==10.4.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
gunicorn==23.0.0
